refactor y limpieza
