+++ tests/Eval.test.ts
@@ -1,6 +1,7 @@
-import { evalGrammar } from "../parser/Eval";
 import { parseSource } from "../parser/Parser";
-import { genGrammar, toExpr } from "../parser/Grammar";
+import { genGrammar } from "../parser/Grammar";
+import { parseToExpr } from "../parser/Expr";
+import { evalGrammar } from "../parser/Eval";
 
 import { TADS, EVALS } from "./Common";
 
@@ -9,13 +10,11 @@ const grammar = genGrammar(tads);
 
 for (const _eval of EVALS) {
     it(`eval ${_eval.line.toString().padStart(3)}:  ${_eval.left.padEnd(40)} -> ${_eval.right.padStart(15)}`, () => {
-        const exprL = toExpr(_eval.left, grammar);
-        const exprR = toExpr(_eval.right, grammar);
+        const exprL = parseToExpr(_eval.left, {}, grammar);
+        const exprR = parseToExpr(_eval.right, {}, grammar);
 
         expect(exprL).not.toBeNull();
         expect(exprR).not.toBeNull();
-
-        // console.log(exprL);
         expect(evalGrammar(exprL!, grammar)).toStrictEqual(exprR!);
     });
 }
