+++ tests/Grammar.test.ts
@@ -1,26 +1,27 @@
 import { parseSource } from "../parser/Parser";
-import { fromExpr, genGrammar, toExpr } from "../parser/Grammar";
+import { genGrammar } from "../parser/Grammar";
 
 import { INVALID_STATEMENTS, VALID_STATEMENTS, TADS } from "./Common";
+import { exprToString, parseToExpr } from "../parser/Expr";
 
 const [tads] = parseSource(TADS.join("\n"));
 
 const grammar = genGrammar(tads);
 const padSize = 60;
 
-it("vacio tiene que fallar", () => expect(toExpr("", grammar)).toBeNull());
+it("vacio tiene que fallar", () => expect(parseToExpr("", {}, grammar)).toBeNull());
 
 for (const stmt of VALID_STATEMENTS) {
-    it("parsea --- " + stmt.padEnd(padSize), () => expect(toExpr(stmt, grammar)).not.toBeNull());
+    it("parsea --- " + stmt.padEnd(padSize), () => expect(parseToExpr(stmt, {}, grammar)).not.toBeNull());
 }
 
 for (const stmt of VALID_STATEMENTS) {
     it("stmt -> Expr -> stmt -> Expr --- " + stmt.padEnd(padSize), () => {
-        const expr1 = toExpr(stmt, grammar);
+        const expr1 = parseToExpr(stmt, {}, grammar);
         expect(expr1).not.toBeNull();
-        const stmt2 = fromExpr(expr1!, grammar);
+        const stmt2 = exprToString(expr1!, grammar);
         expect(stmt2.length).toBeGreaterThan(0);
-        const expr2 = toExpr(stmt2, grammar);
+        const expr2 = parseToExpr(stmt2, {}, grammar);
         expect(expr2).not.toBeNull();
 
         expect(expr1).toStrictEqual(expr2);
@@ -29,7 +30,7 @@ for (const stmt of VALID_STATEMENTS) {
 
 for (const stmt of INVALID_STATEMENTS) {
     it("no debería parsear --- " + stmt.padEnd(padSize), () => {
-        const expr = toExpr(stmt, grammar);
+        const expr = parseToExpr(stmt, {}, grammar);
         expect(expr).toBeNull();
     });
 }
